fix: switch link widget files + add one more widget per width
