+++ src/client/services/mapWidgetsService.tsx
@@ -2,8 +2,8 @@ import { type WidgetData } from "~/server/domain/config/widgetData";
 import { type AdjustedWidgetLayout } from "~/server/domain/layout/adjustedWidgetLayout";
 import { type CalendarWidgetData } from "~/server/widgets/calendar/types";
 import CalendarWidget from "~/server/widgets/calendar/widget";
-import LinkCollectionWidget from "~/server/widgets/links/LinkCollectionWidget";
 import { type LinkWidgetData } from "~/server/widgets/links/types";
+import LinkCollectionWidget from "~/server/widgets/links/widget";
 import { type SearchWidgetData } from "~/server/widgets/search/types";
 import SearchWidget from "~/server/widgets/search/widget";
 import { type TimeWidgetData } from "~/server/widgets/time/types";
